% Generated by roxygen2: do not edit by hand
\name{cluster_sd_to_icc}
\alias{cluster_sd_to_icc}
\title{Convert a random-intercept standard deviation to a logistic-model ICC}
\usage{
cluster_sd_to_icc(cluster_sd)
}
\arguments{
\item{cluster_sd}{Nonnegative random-intercept standard deviation.}
}
\value{The corresponding latent-scale intraclass correlation coefficient.}
\description{Converts a logistic random-intercept standard deviation to an intraclass correlation coefficient.}
