% Generated by roxygen2: do not edit by hand
\name{icc_to_cluster_sd}
\alias{icc_to_cluster_sd}
\title{Convert a logistic-model ICC to a random-intercept standard deviation}
\usage{
icc_to_cluster_sd(icc)
}
\arguments{
\item{icc}{Intraclass correlation coefficient in the open interval (0, 1).}
}
\value{The corresponding random-intercept standard deviation.}
\description{Uses the latent-variable approximation for a logistic random-intercept model, where the level-1 variance is pi^2 / 3.}
