% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{cluster_sd_to_icc}
\alias{cluster_sd_to_icc}
\title{Convert a random-intercept standard deviation to a logistic-model ICC}
\usage{
cluster_sd_to_icc(cluster_sd)
}
\arguments{
\item{cluster_sd}{Nonnegative random-intercept standard deviation.}
}
\value{
The corresponding latent-scale intraclass correlation coefficient.
}
\description{
Convert a random-intercept standard deviation to a logistic-model ICC
}
\examples{
cluster_sd_to_icc(0.416)
}
