% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistage_power.R
\name{sw_multistage_power}
\alias{sw_multistage_power}
\alias{print.sw_multistage_power}
\alias{summary.sw_multistage_power}
\alias{plot.sw_multistage_power}
\title{Methods for multistage stepped-wedge power objects}
\usage{
\method{print}{sw_multistage_power}(x, ...)

\method{summary}{sw_multistage_power}(object, ...)

\method{plot}{sw_multistage_power}(x, type = c("conditional", "failure-aware"), ...)
}
\arguments{
\item{x, object}{An \code{sw_multistage_power} object.}

\item{...}{Additional graphical arguments for \code{plot}; otherwise unused.}

\item{type}{Power definition for the plot: \code{"conditional"} or
\code{"failure-aware"}.}
}
\description{
Methods for multistage stepped-wedge power objects
}
